requirements_dev.txt
-r requirements.txt
ipdb==0.13.9
flake8==4.0.1
pytest==7.1.2
pytest-flask==1.2.0
pytest-cov==3.0.0
pytest-pep8==1.0.6
pytest-watch==4.2.0
pytest-flake8==1.1.1
pytest-dotenv==0.5.2
flask-shell-ipython==0.4.1
vcrpy==4.2.0
mock==4.0.3
freezegun==1.2.2